+++ src/subclient.ts
@@ -618,7 +618,7 @@ export class SubClient {
     price: number,
     size: number,
     side: types.Side,
-    options: types.OrderOptions
+    options: types.OrderOptions = types.defaultOrderOptions()
   ): Promise<TransactionSignature> {
     let tx = new Transaction();
     let marketIndex = this._subExchange.markets.getMarketIndex(market);
@@ -645,8 +645,7 @@ export class SubClient {
 
     let tifOffsetToUse = utils.getTIFOffset(
       this._subExchange.markets.getMarket(market),
-      options.explicitTIF != undefined ? options.explicitTIF : true,
-      options.tifOffset != undefined ? options.tifOffset : 0
+      options.tifOptions
     );
 
     let orderIx = instructions.placeOrderV4Ix(
@@ -700,7 +699,7 @@ export class SubClient {
     price: number,
     size: number,
     side: types.Side,
-    options: types.OrderOptions
+    options: types.OrderOptions = types.defaultOrderOptions()
   ): Promise<TransactionSignature> {
     let tx = new Transaction();
     let market = Exchange.getPerpMarket(this._asset).address;
@@ -728,9 +727,9 @@ export class SubClient {
 
     let tifOffsetToUse = utils.getTIFOffset(
       this._subExchange.markets.getMarket(market),
-      options.explicitTIF != undefined ? options.explicitTIF : true,
-      options.tifOffset != undefined ? options.tifOffset : 0
+      options.tifOptions
     );
+
     let orderIx = instructions.placePerpOrderV2Ix(
       this.asset,
       marketIndex,
@@ -797,7 +796,7 @@ export class SubClient {
     price: number,
     size: number,
     side: types.Side,
-    options: types.OrderOptions
+    options: types.OrderOptions = types.defaultOrderOptions()
   ): TransactionInstruction {
     if (marketIndex == constants.PERP_INDEX) {
       return this.createPlacePerpOrderInstruction(price, size, side, options);
@@ -805,8 +804,7 @@ export class SubClient {
 
     let tifOffsetToUse = utils.getTIFOffset(
       Exchange.getMarket(this._asset, marketIndex),
-      options.explicitTIF != undefined ? options.explicitTIF : true,
-      options.tifOffset != undefined ? options.tifOffset : 0
+      options.tifOptions
     );
 
     return instructions.placeOrderV4Ix(
@@ -832,7 +830,7 @@ export class SubClient {
     price: number,
     size: number,
     side: types.Side,
-    options: types.OrderOptions
+    options: types.OrderOptions = types.defaultOrderOptions()
   ): TransactionInstruction {
     if (
       this._openOrdersAccounts[constants.PERP_INDEX].equals(PublicKey.default)
@@ -844,6 +842,13 @@ export class SubClient {
       );
       throw Error("User does not have an open orders account.");
     }
+
+    let market = Exchange.getPerpMarket(this._asset).address;
+    let tifOffset = utils.getTIFOffset(
+      this._subExchange.markets.getMarket(market),
+      options.tifOptions
+    );
+
     return instructions.placePerpOrderV2Ix(
       this.asset,
       constants.PERP_INDEX,
@@ -855,7 +860,7 @@ export class SubClient {
         : types.OrderType.LIMIT,
       options.clientOrderId != undefined ? options.clientOrderId : 0,
       options.tag != undefined ? options.tag : constants.DEFAULT_ORDER_TAG,
-      options.tifOffset != undefined ? options.tifOffset : 0,
+      tifOffset,
       this.marginAccountAddress,
       this._parent.publicKey,
       this._openOrdersAccounts[constants.PERP_INDEX],
@@ -935,7 +940,7 @@ export class SubClient {
     newOrderPrice: number,
     newOrderSize: number,
     newOrderSide: types.Side,
-    options: types.OrderOptions
+    options: types.OrderOptions = types.defaultOrderOptions()
   ): Promise<TransactionSignature> {
     let tx = new Transaction();
     let marketIndex = this._subExchange.markets.getMarketIndex(market);
@@ -953,8 +958,7 @@ export class SubClient {
 
     let tifOffsetToUse = utils.getTIFOffset(
       this._subExchange.markets.getMarket(market),
-      options.explicitTIF != undefined ? options.explicitTIF : true,
-      options.tifOffset != undefined ? options.tifOffset : 0
+      options.tifOptions
     );
 
     tx.add(
@@ -996,7 +1000,7 @@ export class SubClient {
     newOrderPrice: number,
     newOrderSize: number,
     newOrderSide: types.Side,
-    newOptions: types.OrderOptions
+    newOptions: types.OrderOptions = types.defaultOrderOptions()
   ): Promise<TransactionSignature> {
     let tx = new Transaction();
     let marketIndex = this._subExchange.markets.getMarketIndex(market);
@@ -1013,8 +1017,7 @@ export class SubClient {
 
     let tifOffsetToUse = utils.getTIFOffset(
       this._subExchange.markets.getMarket(market),
-      newOptions.explicitTIF != undefined ? newOptions.explicitTIF : true,
-      newOptions.tifOffset != undefined ? newOptions.tifOffset : 0
+      newOptions.tifOptions
     );
     tx.add(
       instructions.placeOrderV4Ix(
@@ -1058,7 +1061,7 @@ export class SubClient {
     newOrderPrice: number,
     newOrderSize: number,
     newOrderSide: types.Side,
-    newOptions: types.OrderOptions
+    newOptions: types.OrderOptions = types.defaultOrderOptions()
   ): Promise<TransactionSignature> {
     let tx = new Transaction();
     let marketIndex = this._subExchange.markets.getMarketIndex(market);
@@ -1075,9 +1078,9 @@ export class SubClient {
 
     let tifOffsetToUse = utils.getTIFOffset(
       this._subExchange.markets.getMarket(market),
-      newOptions.explicitTIF != undefined ? newOptions.explicitTIF : true,
-      newOptions.tifOffset != undefined ? newOptions.tifOffset : 0
+      newOptions.tifOptions
     );
+
     tx.add(
       instructions.placeOrderV4Ix(
         this.asset,